Enable enterprise features and add performance metrics
Re-enabled previously excluded enterprise services and view models in project files and DI setup. Added new performance-related properties to models and updated service interfaces and implementations to support performance metrics, health score history, and forecasting. Also included PostgreSQL adapter in data project and updated UI to show executive dashboard and compliance audit tabs.

+++ DBOptimizer.Core/Services/QueryPerformanceForecastingService.cs
@@ -81,8 +81,8 @@ public class QueryPerformanceForecastingService : IQueryPerformanceForecastingSe
             {
                 Date = lastTimestamp.AddDays(day),
                 PredictedValue = Math.Max(0, predictedValue),
-                LowerBound = Math.Max(0, predictedValue - (2 * stdDev)),
-                UpperBound = predictedValue + (2 * stdDev),
+                LowerBound = (int)Math.Max(0, predictedValue - (2 * stdDev)),
+                UpperBound = (int)(predictedValue + (2 * stdDev)),
                 IsAnomaly = false
             };
